chore(ErrorFormTooltip): rename folder
