+++ packages/orbit-components/src/ErrorFormTooltip/Tooltip/helpers/__tests__/resolveColor.test.js
