+++ packages/orbit-components/src/ErrorFormTooltip/Tooltip/helpers/resolveColor.js
