+++ packages/orbit-components/src/ErrorFormTooltip/Tooltip/helpers/resolveTooltipArrowPosition.js.flow
