+++ packages/orbit-components/src/ErrorFormTooltip/Tooltip/helpers/resolveTooltipPosition.js.flow
