+++ packages/orbit-components/src/ErrorFormTooltip/Tooltip/helpers/tooltipArrowStyle.js
