+++ packages/orbit-components/src/ErrorFormTooltip/helpers/resolveCustomPosition.js
