+++ packages/orbit-components/src/ErrorFormTooltip/helpers/resolveTooltipPosition.js
