+++ packages/orbit-components/src/InputField/index.js
@@ -10,11 +10,11 @@ import InputTags from "./InputTags";
 import getSpacingToken from "../common/getSpacingToken";
 import getFieldDataState from "../common/getFieldDataState";
 import useRandomId from "../hooks/useRandomId";
-import TooltipForm from "../TooltipForm";
+import TooltipForm from "../ErrorFormTooltip";
 import AlertCircle from "../icons/AlertCircle";
 import InformationCircle from "../icons/InformationCircle";
 import FormLabel from "../FormLabel";
-import useErrorTooltip from "../TooltipForm/hooks/useErrorTooltip";
+import useErrorTooltip from "../ErrorFormTooltip/hooks/useErrorTooltip";
 import formElementFocus from "./helpers/formElementFocus";
 import { StyledButtonPrimitiveIconContainer } from "../primitives/ButtonPrimitive/components/ButtonPrimitiveIconContainer";
 import mq from "../utils/mediaQuery";
